fix import path and rename to `wf`
